[All] Fix watchdog bugs
In the case of "timeout is 3 sec & sleep is 2 sec", the system still goes to reboot because the intentional delays in ping function. The delay also causes kernel panic in restart handler when doing reboot.

In addtion, I2C commands may fail during reboot procedure. To avoid these problems, we remove all the delays, set timeout in shutdown function, and only do 'watchdog enable' in restart handler.

The commit #a88584e5 is reverted as well, and finally we enable external watchdog for all imx6 projects.